refactor: move CustomEmoji and CustomEmojiContent to emoji module and update references
