+++ notionrs/src/object/mod.rs
@@ -1,6 +1,5 @@
 pub mod block;
 pub mod color;
-pub mod custom_emoji;
 pub mod database;
 pub mod emoji;
 pub mod file;
